# docs/swagger.yaml
definitions:
...
  models.Reward:
    properties:
      id:
        type: integer
      name:
        type: string
      project:
        type: string
      type:
        type: string
      winner:
        type: string
      year:
        type: integer
    type: object
  models.Work:
    properties:
      desc:
        type: string
      id:
        type: integer
      type:
        type: string
      year:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/reward:
    post:
      consumes:
      - application/json
      description: 新增奖励（竞赛名称，项目名称，年份，类型，获奖人）
      parameters:
      - description: Add Reward Request
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/models.Reward'
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 500, msg: ''INVALID_PARAMS'', data: nil'
          schema:
            additionalProperties: true
            type: object
      summary: 新增奖励
      tags:
      - 奖励
  /api/admin/reward/{id}:
    delete:
      consumes:
      - application/json
      description: 通过奖励ID删除一个奖励项
      parameters:
      - description: 奖励ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 10002, msg: ''ERROR_NOT_EXIST_REWARD'', data: {}'
          schema:
            additionalProperties: true
            type: object
      summary: 删除奖励
      tags:
      - 奖励
    put:
      consumes:
      - application/json
      description: 根据id更新已存在的奖励
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Reward Request
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/models.Reward'
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 10002, msg: ''ERROR_NOT_EXIST_REWARD'', data: nil'
          schema:
            additionalProperties: true
            type: object
      summary: 编辑奖励
      tags:
      - 奖励
  /api/admin/work:
    post:
      consumes:
      - application/json
      description: 添加一个新的成果记录
      parameters:
      - description: 成果详情
        in: body
        name: work
        required: true
        schema:
          $ref: '#/definitions/models.Work'
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 400, msg: ''INVALID_PARAMS'', data: {}'
          schema:
            additionalProperties: true
            type: object
      summary: 新增成果
      tags:
      - 成果
  /api/admin/work/{id}:
    delete:
      consumes:
      - application/json
      description: 通过成果ID删除一个成果记录
      parameters:
      - description: 成果ID
        in: path
        name: id
        required: true
        type: integer
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 10001, msg: ''ERROR_NOT_EXIST_WORK'', data: {}'
          schema:
            additionalProperties: true
            type: object
      summary: 删除成果
      tags:
      - 成果
    put:
      consumes:
      - application/json
      description: 根据ID更新成果的详细信息
      parameters:
      - description: 成果ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新成果详情
        in: body
        name: work
        required: true
        schema:
          $ref: '#/definitions/models.Work'
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 10001, msg: ''ERROR_NOT_EXIST_WORK'', data: {}'
          schema:
            additionalProperties: true
            type: object
      summary: 更新成果
      tags:
      - 成果
...
  /api/rewards:
    get:
      consumes:
      - application/json
      description: 根据类型和项目名称查询
      parameters:
      - description: 项目名称
        in: query
        name: project
        type: string
      - description: 类型（竞赛/其他）
        in: query
        name: type
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200, msg: ''success'', data: {lists, total}'
          schema:
            additionalProperties: true
            type: object
      summary: 获取奖励列表
      tags:
      - 奖励
...
  /api/works:
    get:
      consumes:
      - application/json
      description: 根据名称和类型进行筛选，获取成果列表
      parameters:
      - description: 成果名称（模糊匹配）
        in: query
        name: name
        type: string
      - description: 成果类型
        in: query
        name: type
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200, msg: ''SUCCESS'', data: {lists: 成果列表, total: 总数}'
          schema:
            additionalProperties: true
            type: object
      summary: 获取成果列表
      tags:
      - 成果
...
swagger: "2.0"